ml-service/requirements.txt
fastapi==0.109.0
uvicorn==0.27.0
python-multipart==0.0.6
pdfplumber==0.10.3
python-dotenv==1.0.0
spacy==3.7.2
google-generativeai==0.3.2
